added actions returncode, minor bugfix in basetheme

+++ .github/actions/build-documentation/entrypoint.sh
@@ -5,6 +5,11 @@ pip install -r requirements.txt
 
 # parse documentation
 python luke.py src/luke/docs/ -o . --resources-with-file -t documentation.html --cdn
+if [ $? -ne 0 ]; then
+    echo "Could not parse Documentation successfully"
+    echo "Exiting..."
+    exit 1
+fi
 
 # remove all markdown files
 find docs -name "*.md" -delete;
